Adding support for sequence learning (LLM)
This changeset adds support for sequence learning features, as they are
used in (large) language models for example. For now it is limited to
generative sequence learning and rather specific to operators used in
the Llama family of LLM models (but can be extended). As we are mainly
targeting smaller GPUs, only 4-bit quantized models using GPTQ quantization
are supported as of now. Other quantization types may follow later.

Also added is a sample application that can run a 7B parameter Llama
model in a simple chat scenario on a 6GB VRAM GPU. Due to lack of
AMD GPUs on the author's side, only (lower end) NVIDIA GPUs were used
for testing (e.g. GTX 1660 Super).

Unfortunately, changes to the high-level interface had to be done in
order to make parameter loading more comfortable. We sincerely hope that
this will be the last high-level API change we had to do. Changes in the
internal API will still follow in order to get rid of all the original
sins that were committed in early development.